# fix: speed up conversation list