--- a/apps/web/src/components/utils/hooks/useMessagePreviews.tsx
+++ b/apps/web/src/components/utils/hooks/useMessagePreviews.tsx
@@ -198,9 +198,13 @@ const useMessagePreviews = () => {
       }
     };
 
-    listConversations();
-    streamConversations();
-    streamAllMessages();
+    const setupConversationsList = async () => {
+      await listConversations();
+      await streamConversations();
+      await streamAllMessages();
+    };
+
+    setupConversationsList();
 
     return () => {
       closeConversationStream();