@neekolas 2022-12-06
I don't understand the part about 
> all awaiting on the same conversations object causing it to lock up

We don't have any explicit mutexes or locks on any objects. Are we sure that it wasn't just doing duplicate HTTP requests or something? The browser will throttle at some number of concurrent http reqs