@elisealix22 2022-12-07
Updated to a WIP / draft PR but would still love to verify that it's significantly faster for Yoginth if possible! It seems specific to `streamAllMessages()` at this point.