@bigint 2022-12-08
@elisealix22 I have the same 😕 

kinda same speed and more requests too

<img width="203" alt="Screenshot 2022-12-08 at 11 52 09 AM" src="https://user-images.githubusercontent.com/69431456/206373006-b3e440e1-ac0d-4244-a525-9a8ffbc18206.png">