@elisealix22 2022-12-08
Thank you so much for testing, Yoginth. Yeah, this wouldn't have removed any `query`s, only two `subscribe`s. Going to spend all of today on this today and then regroup with the team to brainstorm other solutions. So far we've ruled out:

1. Only loading 10 preview messages at a time doesn't resolve the issue -- the bottleneck appears to be in the listing of conversations.
2. It's not a concurrency problem (which is what I had originally thought here). Neither sequencing the `conversations.list()` or removing the duplicate calls speeds up Yoginth's time to see the conversation list.